Added a functioning derivative estimator and polynomial deriver.
Note: Futer work still must be done on polynomials in of themselves and a smart comparer (one that looks at more than just the string, but rather looks and the math as math, like 'x + 1 = 1 + x' are up on the drawing board to be added as
a part of utils. This will make tests and builds cleaner.